Move sass types into sass module.
